+++ crates/binding_core_node/src/util.rs
@@ -1,11 +1,9 @@
-use std::{
-    any::type_name,
-    panic::{catch_unwind, AssertUnwindSafe},
-};
+#![deny(warnings)]
 
-use anyhow::{anyhow, Context, Error};
-use napi::{Env, Status};
-use serde::de::DeserializeOwned;
+use std::panic::{catch_unwind, AssertUnwindSafe};
+
+use anyhow::{anyhow, Error};
+use napi::Env;
 use swc::try_with_handler;
 use swc_common::{
     errors::Handler,
@@ -15,7 +13,7 @@ use swc_common::{
 use tracing::instrument;
 use tracing_chrome::ChromeLayerBuilder;
 use tracing_subscriber::{
-    filter, prelude::__tracing_subscriber_SubscriberExt, util::SubscriberInitExt, EnvFilter, Layer,
+    filter, prelude::__tracing_subscriber_SubscriberExt, util::SubscriberInitExt, Layer,
 };
 
 static TARGET_TRIPLE: &str = include_str!(concat!(env!("OUT_DIR"), "/triple.txt"));
@@ -57,19 +55,6 @@ pub fn init_custom_trace_subscriber(
     Ok(())
 }
 
-/// Trying to initialize default subscriber if global dispatch is not set.
-/// This can be called multiple time, however subsequent calls will be ignored
-/// as tracing_subscriber only allows single global dispatch.
-pub fn init_default_trace_subscriber() {
-    let _unused = tracing_subscriber::FmtSubscriber::builder()
-        .without_time()
-        .with_target(false)
-        .with_ansi(true)
-        .with_env_filter(EnvFilter::from_env("SWC_LOG"))
-        .with_env_filter(EnvFilter::from_default_env().add_directive(tracing::Level::ERROR.into()))
-        .try_init();
-}
-
 #[instrument(level = "trace", skip_all)]
 pub fn try_with<F, Ret>(cm: Lrc<SourceMap>, skip_filename: bool, op: F) -> Result<Ret, Error>
 where
@@ -99,44 +84,4 @@ where
             }
         },
     )
-}
-
-pub trait MapErr<T>: Into<Result<T, anyhow::Error>> {
-    fn convert_err(self) -> napi::Result<T> {
-        self.into()
-            .map_err(|err| napi::Error::new(Status::GenericFailure, format!("{:?}", err)))
-    }
-}
-
-impl<T> MapErr<T> for Result<T, anyhow::Error> {}
-
-pub(crate) fn get_deserialized<T, B>(buffer: B) -> napi::Result<T>
-where
-    T: DeserializeOwned,
-    B: AsRef<[u8]>,
-{
-    let mut deserializer = serde_json::Deserializer::from_slice(buffer.as_ref());
-    deserializer.disable_recursion_limit();
-
-    let v = T::deserialize(&mut deserializer)
-        .with_context(|| {
-            format!(
-                "Failed to deserialize buffer as {}\nJSON: {}",
-                type_name::<T>(),
-                String::from_utf8_lossy(buffer.as_ref())
-            )
-        })
-        .convert_err()?;
-
-    Ok(v)
-}
-
-pub(crate) fn deserialize_json<T>(json: &str) -> Result<T, serde_json::Error>
-where
-    T: DeserializeOwned,
-{
-    let mut deserializer = serde_json::Deserializer::from_str(json);
-    deserializer.disable_recursion_limit();
-
-    T::deserialize(&mut deserializer)
 }