+++ crates/swc_error_reporters/src/lib.rs
@@ -13,6 +13,8 @@ use swc_common::{
     BytePos, FileName, SourceMap, Span,
 };
 
+pub mod handler;
+
 pub struct PrettyEmitter {
     cm: Lrc<SourceMap>,